% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.r
\name{mr_copy}
\alias{mr_copy}
\title{MR Copy}
\usage{
mr_copy(mr)
}
\arguments{
\item{mr}{An 'mr' object, as returned from \code{mr_create()}.}
}
\value{
An object of class \code{MR} (an external pointer).
}
\description{
MR Copy
}

