% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.r
\name{mr_create}
\alias{mr_create}
\title{MR Create}
\usage{
mr_create(mapstyle = "chunk", verbosity = "none", timer = "none",
  comm = 0)
}
\arguments{
\item{mapstyle}{TODO}

\item{verbosity}{TODO}

\item{timer}{TODO}

\item{comm}{MPI communicator number.}
}
\value{
An object of class \code{MR} (an external pointer).
}
\description{
MR Create
}

