% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbdMR-package.R
\docType{package}
\name{pbdMR-package}
\alias{pbdMR-package}
\title{MapReduce}
\description{
An interface to MapReduce over MPI.
}
\details{
This package requires an MPI library (OpenMPI, MPICH2, or LAM/MPI).
}
\author{
Drew Schmidt, Wei-Chen Chen, George Ostrouchov.
}
\references{
Programming with Big Data in R Website: \url{http://r-pbd.org/}
}
\keyword{Package}

